requirements.txt
langchain-community
langchain-openai
PyPDF2
python-docx
qdrant-client
langchain-qdrant
fastapi
uvicorn
torch
sentence-transformers
langchain-huggingface
huggingface-hub
python-multipart
python-dotenv
google-generativeai
langchain-google-genai